Calculate EQ volume after post-download instead of pre-play
