+++ musicbot/player.py
@@ -239,57 +239,6 @@ class MusicPlayer(EventEmitter, Serializable):
     def play(self, _continue=False):
         self.loop.create_task(self._play(_continue=_continue))
 
-    async def run_command(self, cmd):
-        p = await asyncio.create_subprocess_shell(cmd, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE)
-        log.debug('Starting asyncio subprocess ({0}) with command: {1}'.format(p, cmd))
-        stdout, stderr = await p.communicate()
-        return stdout + stderr
-
-    def get(self, program):
-        def is_exe(fpath):
-            found = os.path.isfile(fpath) and os.access(fpath, os.X_OK)
-            if not found and sys.platform == 'win32':
-                fpath = fpath + ".exe"
-                found = os.path.isfile(fpath) and os.access(fpath, os.X_OK)
-            return found
-
-        fpath, __ = os.path.split(program)
-        if fpath:
-            if is_exe(program):
-                return program
-        else:
-            for path in os.environ["PATH"].split(os.pathsep):
-                path = path.strip('"')
-                exe_file = os.path.join(path, program)
-                if is_exe(exe_file):
-                    return exe_file
-
-        return None
-
-    async def get_mean_volume(self, input_file):
-        log.debug('Calculating mean volume of {0}'.format(input_file))
-        try:
-            cmd = '"' + self.get('ffmpeg') + '" -i "' + input_file + '" -af "volumedetect" -f null /dev/null'
-            output = await self.run_command(cmd)
-        except Exception as e:
-            raise e
-        output = output.decode("utf-8")
-        # print('----', output)
-        mean_volume_matches = re.findall(r"mean_volume: ([\-\d\.]+) dB", output)
-        if (mean_volume_matches):
-            mean_volume = float(mean_volume_matches[0])
-        else:
-            mean_volume = float(0)
-
-        max_volume_matches = re.findall(r"max_volume: ([\-\d\.]+) dB", output)
-        if (max_volume_matches):
-            max_volume = float(max_volume_matches[0])
-        else:
-            max_volume = float(0)
-
-        log.debug('Calculated mean volume as {0}'.format(mean_volume))
-        return mean_volume, max_volume
-
     async def _play(self, _continue=False):
         """
             Plays the next entry from the playlist, or resumes playback of the current entry if paused.
@@ -319,18 +268,11 @@ class MusicPlayer(EventEmitter, Serializable):
 
                 boptions = "-nostdin"
                 # aoptions = "-vn -b:a 192k"
-                if self.bot.config.use_experimental_equalization and not isinstance(entry, StreamPlaylistEntry):
-                    try:
-                        mean, maximum = await self.get_mean_volume(entry.filename)
-                        aoptions = '-af "volume={}dB"'.format((maximum * -1))
-                    except Exception as e:
-                        raise
-                        #log.warning('There was a problem with EQ. The track will not be equalised. For full info, check your log file.')
-                        #log.debug('{0}'.format(e))
-                        aoptions = "-vn"
+                if isinstance(entry, StreamPlaylistEntry):
+                    aoptions = entry.aoptions
                 else:
                     aoptions = "-vn"
-                
+
                 log.ffmpeg("Creating player with options: {} {} {}".format(boptions, aoptions, entry.filename))
 
                 self._current_player = self._monkeypatch_player(self.voice_client.create_ffmpeg_player(
